code review: move setup_bitgen inside _random.cc and remove useless random_utils.h
