+++ src/python/gudhi/datasets/generators/_points.cc
@@ -23,7 +23,6 @@
 
 #include <gudhi/Debug_utils.h>
 #include <python_interfaces/numpy_utils.h>
-#include <python_interfaces/random_utils.h>
 
 namespace nb = nanobind;
